Move login functionnilities to subfolder
