+++ frontend/src/components/login/login.js
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { Link, Redirect } from "react-router-dom";
 import axios from 'axios';
 import { Card, Form, Input, Button, Error } from "./AuthForm";
-import { useAuth } from "./context/auth";
-import './Modal.css'
+import { useAuth } from "../context/auth";
+import '../Modal.css'
 
 function Login({toggleModalBox, setDisplayModal}) {
   const [isLoggedIn, setLoggedIn] = useState(false);
